Rename EventCard to SlideCard
Also removed the duplicate button
