+++ src/components/SlideCard.vue
